# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "persisted_revision": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "NetworkSettings": {
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "persistent_keepalive": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "subnet": {
            "type": "string"
          }
        },
        "required": [
          "subnet",
          "endpoint",
          "persistent_keepalive"
        ],
        "type": "object"
      },
...
      "SettingsRequest": {
        "additionalProperties": false,
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "persistent_keepalive": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "endpoint",
          "persistent_keepalive"
        ],
        "type": "object"
      },
      "SetupRequest": {
        "additionalProperties": false,
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "persistent_keepalive": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "subnet": {
            "type": "string"
          }
        },
        "required": [
          "subnet",
          "endpoint",
          "persistent_keepalive"
        ],
        "type": "object"
      },
      "SetupStatus": {
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NetworkSettings"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "configured"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminBearer": {
        "description": "Exact configured WIREHUB_ADMIN_TOKEN. Keep it outside URLs and logs.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "wirehub",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/settings": {
      "put": {
        "operationId": "put_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkSettings"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
    "/api/setup": {
      "get": {
        "operationId": "get_setup",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      },
      "post": {
        "operationId": "post_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkSettings"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
...
  }
}